[FIX] Não perder sequencias ao atualizar hr_payroll

+++ l10n_br_hr_payroll/models/hr_salary_rule.py
@@ -36,7 +36,7 @@ CALCULO_FOLHA_PT_BR = {
 class HrSalaryRule(models.Model):
     _inherit = 'hr.salary.rule'
 
-    sequence = fields.Float(
+    sequence = fields.Integer(
         string=u'Sequence',
     )
     compoe_base_INSS = fields.Boolean(
